Modularity Fix
Modularity Fix for Facebook and Twitter
