+++ app/src/main/AndroidManifest.xml
@@ -24,6 +24,14 @@
         </activity>
 
         <activity
+            android:name=".activities.TwitterLoginActivity"
+            android:label="@string/app_name"
+            android:screenOrientation="portrait" />
+        <activity
+            android:name=".activities.FacebookLoginActivity"
+            android:label="@string/app_name"
+            android:screenOrientation="portrait" />
+        <activity
             android:name=".adapter.FragmentAdapterChooser"
             android:label="@string/app_name"
             android:screenOrientation="portrait" />
