+++ app/src/main/java/hr/mars/muzicow/adapter/FragmentAdapterChooser.java
@@ -167,6 +167,8 @@ public class FragmentAdapterChooser extends AppCompatActivity {
     public void logout() {
         Intent intent = new Intent(this, MainActivity.class);
         startActivity(intent);
+
     }
+
 
 }
