+++ app/src/main/res/layout/social_network_login.xml
@@ -1,62 +1,50 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:gravity="center_horizontal"
     android:orientation="vertical">
 
-    <com.twitter.sdk.android.core.identity.TwitterLoginButton
-        android:id="@+id/twitter"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_above="@+id/login_buttonFB"
-        android:layout_alignEnd="@+id/login_buttonFB"
-        android:layout_alignStart="@+id/login_buttonFB"
-        android:layout_marginBottom="29dp"
-        android:text="Login with Twitter" />
-
-    <com.facebook.login.widget.LoginButton
-        android:id="@+id/login_buttonFB"
-        android:layout_width="259dp"
-        android:layout_height="200dp"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginBottom="37dp"
-        android:paddingBottom="13dp"
-        android:paddingLeft="20dp"
-        android:paddingTop="13dp"
-        android:scaleType="centerInside"
-        android:textSize="18dp" />
-
-    <Spinner
-        android:id="@+id/spinner"
-        style="@style/Base.Animation.AppCompat.DropDownUp"
-        android:layout_width="130dp"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/chooseTxt"
-        android:layout_centerHorizontal="true"
-        android:layout_gravity="center_horizontal"
-        android:entries="@array/spinner"
-        android:spinnerMode="dialog" />
-
-    <TextView
-        android:id="@+id/chooseTxt"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignEnd="@+id/imageView2"
-        android:layout_alignParentTop="true"
-        android:layout_alignStart="@+id/imageView2"
-        android:layout_marginTop="20dp"
-        android:gravity="center_horizontal"
-        android:text="Choose your role"
-        android:textAppearance="?android:attr/textAppearanceMedium" />
 
     <ImageView
         android:id="@+id/imageView2"
         android:layout_width="200dp"
         android:layout_height="200dp"
-        android:layout_centerHorizontal="true"
-        android:layout_centerVertical="true"
-        android:background="@drawable/logo" />
-</RelativeLayout>
+        android:background="@drawable/logo"
+        android:layout_below="@+id/sNChoose"
+        android:layout_alignStart="@+id/loginID" />
+
+    <TextView
+        android:id="@+id/textView3"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:gravity="center_horizontal"
+        android:text="Choose login"
+        android:textAppearance="?android:attr/textAppearanceMedium"
+        android:layout_alignParentTop="true"
+        android:layout_centerHorizontal="true" />
+
+    <Spinner
+        style="@style/Base.Animation.AppCompat.DropDownUp"
+        android:id="@+id/sNChoose"
+        android:layout_width="130dp"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center_horizontal"
+        android:entries="@array/sNChoose"
+        android:spinnerMode="dialog"
+        android:layout_below="@+id/textView3"
+        android:layout_alignStart="@+id/textView3" />
+
+    <Button
+        android:layout_width="241dp"
+        android:layout_height="wrap_content"
+        android:text="Accept"
+        android:id="@+id/loginID"
+        style="?android:attr/buttonStyleSmall"
+        android:background="@drawable/tw__login_btn"
+        android:textColor="#FFFFFF"
+        android:textSize="20sp"
+        android:layout_below="@+id/imageView2"
+        android:layout_alignParentEnd="true"
+        android:layout_alignStart="@+id/sNChoose" />
+</LinearLayout>
