+++ app/src/main/res/layout/twitter_login.xml
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:gravity="center_horizontal"
+    android:orientation="vertical">
+
+
+    <ImageView
+        android:id="@+id/imageView2"
+        android:layout_width="200dp"
+        android:layout_height="200dp"
+        android:background="@drawable/logo"
+        android:layout_below="@+id/spinner"
+        android:layout_alignStart="@+id/twitter" />
+
+    <Spinner
+        android:id="@+id/spinner"
+        style="@style/Base.Animation.AppCompat.DropDownUp"
+        android:layout_width="130dp"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center_horizontal"
+        android:entries="@array/spinner"
+        android:spinnerMode="dialog"
+        android:layout_below="@+id/chooseTxt"
+        android:layout_centerHorizontal="true" />
+
+    <TextView
+        android:id="@+id/chooseTxt"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignEnd="@+id/imageView2"
+        android:layout_alignParentTop="true"
+        android:layout_alignStart="@+id/imageView2"
+        android:layout_marginTop="20dp"
+        android:gravity="center_horizontal"
+        android:text="Choose your role"
+        android:textAppearance="?android:attr/textAppearanceMedium" />
+
+    <com.twitter.sdk.android.core.identity.TwitterLoginButton
+        android:id="@+id/twitter"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Login with Twitter"
+        android:layout_below="@+id/imageView2"
+        android:layout_alignEnd="@+id/spinner" />
+
+
+</LinearLayout>
