+++ app/src/main/res/values/strings.xml
@@ -8,4 +8,9 @@
         <item>Participant</item>
         <item>Artist</item>
     </string-array>
+
+    <string-array name="sNChoose">
+        <item>Twitter</item>
+        <item>Facebook</item>
+    </string-array>
 </resources>
